+++ CHANGELOG.md
@@ -5,7 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2019-05-27
+## Unreleased - 2019-07-15
+### Changed
+- :warning: [Breaking] Metric `mc_players_total` no longer has a `state` label. It exports only the number of unique players now. 
+  Use `mc_players_online_total` for online player count.
+
+### Added
+- Metrics can be enabled individually via config `enable_metrics`. Since the format of the config.yml has changed it is advised to delete your existing file to 
+  generate a new clean one. 
+
+## [v1.3.0] - 2019-05-28
 ### Added
 - This changelog
 - Add `host` config property
@@ -35,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial exporter
 
-[Unreleased]: https://github.com/sladkoff/minecraft-prometheus-exporter/compare/v1.2.0...HEAD
+[v1.3.0]: https://github.com/sladkoff/minecraft-prometheus-exporter/compare/v1.2.0...v1.3.0
 [v1.2.0]: https://github.com/sladkoff/minecraft-prometheus-exporter/compare/v1.1.0...v1.2.0
 [v1.1.0]: https://github.com/sladkoff/minecraft-prometheus-exporter/compare/v1.0.1...v1.1.0
 [v1.0.1]: https://github.com/sladkoff/minecraft-prometheus-exporter/compare/v1.0.0...v1.0.1